# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API V1",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/personajes": {
            "post": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Crear un nuevo personaje",
                "description": "Crea un personaje con un nombre, vida, porcentaje de fallos y habilidades.",
                "operationId": "541defa3a880d2a2e3c2d7d245c80be8",
                "requestBody": {
                    "description": "Datos necesarios para crear un personaje",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "vida",
                                    "miss_percent",
                                    "habilidades"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del personaje",
                                        "type": "string",
                                        "example": "Ezra Thompson"
                                    },
                                    "vida": {
                                        "description": "Cantidad de vida del personaje",
                                        "type": "integer",
                                        "example": 150
                                    },
                                    "miss_percent": {
                                        "description": "Porcentaje de fallo en ataques del personaje",
                                        "type": "integer",
                                        "example": 28
                                    },
                                    "habilidades": {
                                        "description": "Objeto con las habilidades del personaje",
                                        "type": "object",
                                        "example": {
                                            "punio_derecho": 45,
                                            "patada_derecha": 60,
                                            "ataque_especial": 55,
                                            "punio_izquierdo": 39,
                                            "patada_izquierda": 34
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personaje creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "personaje": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Ezra Thompson"
                                                },
                                                "vida": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "miss_percent": {
                                                    "type": "integer",
                                                    "example": 28
                                                },
                                                "habilidades": {
                                                    "type": "object",
                                                    "example": {
                                                        "punio_derecho": 45,
                                                        "patada_derecha": 60,
                                                        "ataque_especial": 55,
                                                        "punio_izquierdo": 39,
                                                        "patada_izquierda": 34
                                                    }
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-11-09 23:53:57"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-11-10 00:48:33"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de entrada no válidos"
                    }
                }
            }
        },
        "/api/v1/salas/{uuid}/personajes": {
            "post": {
                "tags": [
                    "Salas",
                    "Personajes"
                ],
                "summary": "Asignar un personaje a una sala",
                "description": "Este endpoint asigna un personaje y un jugador a una sala, y guarda los valores de vida y porcentaje de fallo del personaje.",
                "operationId": "9815f320f44c553b1b46c8c51dbb2d5b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID de la sala",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos necesarios para asignar un personaje a una sala",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sala_id",
                                    "personaje_id",
                                    "jugador_id"
                                ],
                                "properties": {
                                    "sala_id": {
                                        "description": "ID de la sala",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "personaje_id": {
                                        "description": "ID del personaje a asignar",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "jugador_id": {
                                        "description": "ID del jugador que será asignado al personaje",
                                        "type": "integer",
                                        "example": 7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Personaje y jugador asignados correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Personaje y jugador asociados correctamente a la sala."
                                        },
                                        "sala": {
                                            "$ref": "#/components/schemas/Sala"
                                        },
                                        "personaje": {
                                            "$ref": "#/components/schemas/Personaje"
                                        },
                                        "jugador": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en los datos de entrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan.perez@correo.com"
                    }
                },
                "type": "object"
            },
            "Personaje": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Guerrero"
                    },
                    "vida": {
                        "type": "integer",
                        "maximum": 150,
                        "minimum": 130,
                        "example": 140
                    },
                    "ataque": {
                        "type": "integer",
                        "maximum": 45,
                        "minimum": 25,
                        "example": 35
                    },
                    "defensa": {
                        "type": "integer",
                        "maximum": 50,
                        "minimum": 20,
                        "example": 30
                    },
                    "velocidad": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 70,
                        "example": 85
                    },
                    "habilidades": {
                        "properties": {
                            "puño_derecho": {
                                "type": "integer",
                                "maximum": 50,
                                "minimum": 20,
                                "example": 40
                            },
                            "puño_izquierdo": {
                                "type": "integer",
                                "maximum": 50,
                                "minimum": 20,
                                "example": 35
                            },
                            "patada_derecha": {
                                "type": "integer",
                                "maximum": 50,
                                "minimum": 20,
                                "example": 45
                            },
                            "patada_izquierda": {
                                "type": "integer",
                                "maximum": 50,
                                "minimum": 20,
                                "example": 30
                            },
                            "ataque_especial": {
                                "type": "integer",
                                "maximum": 80,
                                "minimum": 30,
                                "example": 60
                            }
                        },
                        "type": "object"
                    },
                    "creado_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-30T12:00:00Z"
                    },
                    "actualizado_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-30T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Sala": {
                "required": [
                    "uuid",
                    "jugador1_id",
                    "estado"
                ],
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6"
                    },
                    "jugador1_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "jugador2_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "jugador1_personaje_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "jugador2_personaje_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "estado": {
                        "type": "string",
                        "example": "abierta"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-30T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-30T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}